+++ src/main/java/web/mvc/repository/StockRepository.java
@@ -6,7 +6,7 @@ import web.mvc.domain.Stock;
 
 import java.util.List;
 
-public interface StockRepository extends JpaRepository<Stock, Integer> {
+public interface StockRepository extends JpaRepository<Stock, Long> {
 
     @Query("select distinct s from Stock s left join fetch s.product left join fetch s.stockGrade left join fetch s.stockOrganic left join fetch s.farmerUser left join fetch s.userBuyDetails where s.farmerUser.user_seq = ?1")
     List<Stock> findStocksById(long farmerId);
