+++ src/main/java/web/mvc/repository/UserBuyRepository.java
@@ -3,5 +3,5 @@ package web.mvc.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import web.mvc.domain.UserBuy;
 
-public interface UserBuyRepository extends JpaRepository<UserBuy,Integer> {
+public interface UserBuyRepository extends JpaRepository<UserBuy, Long> {
 }
